Add servo control python script
vrcam.py is a simple python script which takes allows three servos to
be moved.  This file currently takes user input to decide the angle for
testing purposes and will later recieve its angles from the mobile
application running on the gear vr.

The pin setup currently uses pins 21, 22 and 23