Add arduino file to control hardware
The raspberry pi has been replaced with a laptop and an arduino due to
the computational limitations of the pi and ease of use of the arduino.
This file controls 2 servos and a stepper motor using a sparkfun
autodriver.

In order to send commands a serial connection between the arduino and
laptop at a baud of 115200 is necessary.  The commands should follow
the structure "yaw:pitch:roll" where yaw, pitch and roll are integers.