Add control file for laptop
The control file for the laptop has been added.  This currently uses 3
threads to accomplish different goals.

The main thread sends serial commands to the arduino thus moving the
servos and the stepper motor.

The camThread currently retrieves and displays an image from the
webcam.  This will be expanded later to allow the camera to transmit to
the android portion of the code.

The sockThread creates a socket server which currently serves forever.
This will also be expanded to retrieve information about yaw pitch and
roll from the android phone.