Add an mjpg stream to the videoThread
Previously the videoThread simply displayed the video on a seperate
window.  This has been modified to allow it to stream over a server
which is currently been set to the local ip 127.0.0.1.  The exact
address of the current server is http://127.0.0.1:8080/cam.mjpg.

This can be viewed using vlc and will later be viewed from the android
device.